+++ apps/backend/src/modules/groups/libs/validation-schemas/validation-schemas.ts
@@ -1,1 +1,6 @@
-export { groupIdParameter } from "@trackmates/shared";
+export {
+	groupIdAndPermissionIdParametersValidationSchema,
+	groupIdAndUserIdParametersValidationSchema,
+	groupIdParameterValidationSchema,
+	groupRequestBodyValidationSchema,
+} from "@trackmates/shared";