+++ apps/frontend/src/libs/constants/menu-items.constants.ts
@@ -1,4 +1,4 @@
-import { AppRoute } from "~/libs/enums/enums.js";
+import { AppRoute, PermissionKey, PermissionMode } from "~/libs/enums/enums.js";
 
 import { type MenuItem } from "../types/types.js";
 
@@ -17,6 +17,15 @@ const MENU_ITEMS: MenuItem[] = [
 		href: AppRoute.FRIENDS,
 		icon: "pie",
 		label: "Friends",
+	},
+	{
+		href: AppRoute.MANAGEMENT,
+		icon: "lock",
+		label: "Management",
+		pagePermissions: {
+			mode: PermissionMode.ONE_OF,
+			permissions: [PermissionKey.MANAGE_UAM, PermissionKey.MANAGE_USERS],
+		},
 	},
 ];