+++ apps/frontend/src/libs/enums/app-title.enum.ts
@@ -3,6 +3,7 @@ const AppTitle = {
 	CHATS: "Chats",
 	COMPARE_PROGRESS: "Compare Progress",
 	FRIENDS: "Friends",
+	MANAGEMENT: "Management",
 	NOTIFICATIONS: "Notifications",
 	PROFILE: "My Profile",
 	SIGN_IN: "Sign In",