+++ apps/frontend/src/libs/enums/enums.ts
@@ -14,5 +14,7 @@ export {
 	NotificationStatus,
 	NotificationType,
 	PaginationValue,
+	PermissionKey,
+	PermissionMode,
 	ServerErrorType,
 } from "@trackmates/shared";