+++ apps/frontend/src/modules/permissions/libs/types/types.ts
@@ -1,0 +1,4 @@
+export {
+	type PermissionResponseDto,
+	type PermissionsGetAllResponseDto,
+} from "@trackmates/shared";