+++ apps/frontend/src/modules/users/libs/types/types.ts
@@ -1,5 +1,6 @@
 export {
 	type UserAuthResponseDto,
+	type UserGetAllResponseDto,
 	type UserProfileRequestDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,