+++ apps/frontend/src/pages/management/libs/components/groups-tab/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { GroupsTableHeader } from "./groups-table-header.enum.js";