+++ apps/frontend/src/pages/management/libs/components/groups-tab/libs/enums/groups-table-header.enum.ts
@@ -1,0 +1,8 @@
+const GroupsTableHeader = {
+	BUTTONS: "",
+	ID: "ID",
+	NAME: "Group name",
+	PERMISSIONS: "Permissions",
+} as const;
+
+export { GroupsTableHeader };