+++ apps/frontend/src/pages/management/libs/components/users-tab/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { UsersTableHeader } from "./users-table-header.enum.js";