+++ packages/shared/src/modules/chat-messages/libs/types/read-chat-messages-response-dto.type.ts
@@ -1,0 +1,8 @@
+import { type ChatMessageItemResponseDto } from "./chat-message-item-response-dto.type.js";
+
+type ReadChatMessagesResponseDto = {
+	items: ChatMessageItemResponseDto[];
+	unreadMessagesCount: number;
+};
+
+export { type ReadChatMessagesResponseDto };